backend/requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.28
alembic==1.13.1
pydantic==2.6.4
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
